requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
tqdm
python-dotenv
pandas